gl2_gears: add the gears demo port that kitty_gears is derived from
kitty_gears is derived from a modified version of the GLFW port
of the public domain OpenGL gears demo. This port adds the gl2_util.h
header with support for loading shaders and creating vertex buffers
and index buffers.

The original version of the gears demo uses the OpenGL 1.x style
fixed-function lighting and immediate mode vertices and call lists.