gl4_gears: experimental port of gears for OpenGL 4.x (#1)
- gl4_gears derived from gl3_gears
  - implemented uniform buffer objects for projection matrices
  - implemented SPIR-V binary shaders
  - enable by changing USE_SPIRV define at top of gl4_gears.c
- gl2_utils OpenGL4.x SPIR-V loading and reflection
  - change link_program to not attempt to rebind attribute
    locations and simply fetch attribute names and their
    default location as reported by glGetAttribLocation.
  - add OpenGL 4.x code to use getProcAddress for new functions
  - use OpenGL 4.x reflection interfaces if available
- update vk1_gears to use the new attribute locations
- add glad to GLFW_LIBS_ALL

there are remaining issues with attribute reflection that need
to be resolved. NVIDIA OpenGL 450 implementation has bugs with
reassignment of attribute locations when using the #version 450
shader dialect used by the SPIR-V shader modules, so the code
has been changed to simply reflect for attribute and uniform
names and accept the default indices provided by the driver.